+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/09 15:13:57 by miyuu             #+#    #+#             */
-/*   Updated: 2025/05/26 18:03:22 by miyuu            ###   ########.fr       */
+/*   Updated: 2025/05/28 00:16:56 by miyuu            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -107,6 +107,7 @@ void			fill_images(t_data *data, const t_tokens_tmp *tokens);
 void			fill_color(t_data *data, const t_tokens_tmp *tokens);
 void			fill_player_position(t_data *data, char **map_lines);
 void			free_data(t_data *d);
+bool			is_valid_map(t_data *data);
 /* init/utils */
 void			str_lstadd_back(t_strlst **lst, t_strlst *new);
 t_strlst		*str_lstnew(char *str);
@@ -119,6 +120,7 @@ bool			is_empty_line(char *line);
 char			*trim_spaces_both_ends(const char *str);
 size_t			get_trimmed_lstsize(const t_strlst *lst);
 char			*strdup_trim_nl(const char *str);
+size_t			calc_map_index(size_t x, size_t y, unsigned int width);
 
 /* dedug 最終的には削除する?*/
 void			debug_print_data(t_data *data);
