+++ src/load/is_valid_map.c
@@ -1,0 +1,83 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   is_valid_map.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/05/28 00:15:40 by miyuu             #+#    #+#             */
+/*   Updated: 2025/05/28 00:24:09 by miyuu            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3d.h"
+
+bool	is_valid_position(t_data *data, size_t x, size_t y)
+{
+	size_t			index;
+	int				*map;
+	unsigned int	width;
+	unsigned int	height;
+
+	map = data->map;
+	width = data->width;
+	height = data->height;
+	/* マップの端にいたらout */
+	if (x == 0 || x == width - 1 || y == 0 || y == height - 1)
+		return (false);
+	/* 現在地が"NOTHING"だったらout */
+	index = calc_map_index(x, y, width);
+	if (map[index] == NOTHING)
+		return (false);
+	return (true);
+}
+
+bool	is_surrounded_walls(t_data *data, size_t x, size_t y, bool *visited)
+{
+	size_t			index;
+	unsigned int	width;
+	int				*map;
+
+	map = data->map;
+	width = data->width;
+	if (!is_valid_position(data, x, y))
+		return (false);
+
+	index = calc_map_index(x, y, width);
+	if (visited[index])
+		return (true);
+	visited[index] = true;
+	/*
+	左右上下に1マス進んだ先を再帰的に検証。
+	→1マス先が壁ではない かつ 訪れたことがない場合、再帰 → 再帰の結果がfalseだったらfalse返す
+	*/
+	//左に1マス
+	if (map[index - 1] != WALL && !visited[index - 1] \
+		&& !is_surrounded_walls(data, x - 1, y, visited))
+		return (false);
+	//右に1マス
+	if (map[index + 1] != WALL && !visited[index + 1] \
+		&& !is_surrounded_walls(data, x + 1, y, visited))
+		return (false);
+	//上に1マス
+	if (map[index - width] != WALL && !visited[index - width] \
+		&& !is_surrounded_walls(data, x, y - 1, visited))
+		return (false);
+	//下に1マス
+	if (map[index + width] != WALL && !visited[index + width] \
+		&& !is_surrounded_walls(data, x, y + 1, visited))
+		return (false);
+	return (true);
+}
+
+bool	is_valid_map(t_data *data)
+{
+	bool	*visited;
+	bool	result;
+
+	visited = (bool *)ft_calloc((data->width * data->height + 1), sizeof(bool));
+	if (!visited)
+		error_perror_and_exit(NULL);
+	result = is_surrounded_walls(data, data->player.x, data->player.y, visited);
+	return (result);
+}
