+++ src/load/utils/calc_map_index.c
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   calc_map_index.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: miyuu <miyuu@student.42.fr>                +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/05/28 00:17:08 by miyuu             #+#    #+#             */
+/*   Updated: 2025/05/28 00:17:14 by miyuu            ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3d.h"
+
+size_t	calc_map_index(size_t x, size_t y, unsigned int width)
+{
+	return (y * width + x);
+}
